Targets sandbox instance of Equipment Facade API
Our example always dealt with sandbox environment. Recently we updated
`fuse-equipment-api.herokuapp.com/` to point to
`agco-fuse-trackers.herokuapp.com`. This means that now, the example is
pointing to production environment.

This change made the example stop working for sandbox users which do not
have production access. In order to fix this problem, this commit
switchs back to sandbox environment.